+++ src/test/java/com/example/demo/service/auth/JwtTokenManagementServiceTest.java
@@ -6,6 +6,7 @@ import com.example.demo.exception.JwtCustomException;
 import com.example.demo.factory.TestJwtFactory;
 import com.example.demo.helper.MemberSaveHelper;
 import com.example.demo.implement.member.Member;
+import com.example.demo.service.token.JwtTokenManagementService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
